Change dev database from sqlite to postgresql.
Since the production database is postgresql I changed the dev database to match so that it is closer to production.  Ths was a good move since there were failures when strings were too long.

The data that is being migrated to this database had some long fields that should have been in the notes column which is a text field. Instead, the data was put into the "vintage" or "winemaker" field which was too long for a string. The error was found early on because I switched to use postgresql for both dev and production.